Made fetching and decrypting maps notes operate in parallel. Fetching can recover from network failures
