+++ crates/sdk/src/masp/mod.rs
@@ -1,0 +1,1434 @@
+//! MASP verification wrappers.
+
+pub mod shielded_ctx;
+pub mod types;
+pub mod utils;
+
+use std::collections::HashMap;
+use std::env;
+use std::fmt::Debug;
+use std::ops::Deref;
+use std::path::PathBuf;
+
+use borsh::{BorshDeserialize, BorshSerialize};
+use lazy_static::lazy_static;
+use masp_primitives::asset_type::AssetType;
+#[cfg(feature = "mainnet")]
+use masp_primitives::consensus::MainNetwork;
+#[cfg(not(feature = "mainnet"))]
+use masp_primitives::consensus::TestNetwork;
+use masp_primitives::convert::AllowedConversion;
+use masp_primitives::ff::PrimeField;
+use masp_primitives::group::GroupEncoding;
+use masp_primitives::memo::MemoBytes;
+use masp_primitives::merkle_tree::MerklePath;
+use masp_primitives::sapling::note_encryption::*;
+use masp_primitives::sapling::redjubjub::PublicKey;
+use masp_primitives::sapling::{Diversifier, Node, Note};
+use masp_primitives::transaction::components::transparent::builder::TransparentBuilder;
+use masp_primitives::transaction::components::{
+    ConvertDescription, I128Sum, OutputDescription, SpendDescription, TxOut,
+    U64Sum,
+};
+use masp_primitives::transaction::fees::fixed::FeeRule;
+use masp_primitives::transaction::sighash::{signature_hash, SignableInput};
+use masp_primitives::transaction::txid::TxIdDigester;
+use masp_primitives::transaction::{
+    Authorization, Authorized, Transaction, TransactionData, TransparentAddress,
+};
+use masp_primitives::zip32::{ExtendedFullViewingKey, ExtendedSpendingKey};
+use masp_proofs::bellman::groth16::PreparedVerifyingKey;
+use masp_proofs::bls12_381::Bls12;
+use masp_proofs::prover::LocalTxProver;
+#[cfg(not(feature = "testing"))]
+use masp_proofs::sapling::SaplingVerificationContext;
+pub use namada_core::masp::{
+    encode_asset_type, AssetData, BalanceOwner, ExtendedViewingKey,
+    PaymentAddress, TransferSource, TransferTarget,
+};
+use namada_token::MaspDigitPos;
+pub use shielded_ctx::ShieldedContext;
+pub use utils::ShieldedUtils;
+
+use crate::masp::types::{PVKs, PartialAuthorized};
+use crate::masp::utils::{get_params_dir, load_pvks};
+use crate::{MaybeSend, MaybeSync};
+
+/// Env var to point to a dir with MASP parameters. When not specified,
+/// the default OS specific path is used.
+pub const ENV_VAR_MASP_PARAMS_DIR: &str = "NAMADA_MASP_PARAMS_DIR";
+
+/// Randomness seed for MASP integration tests to build proofs with
+/// deterministic rng.
+pub const ENV_VAR_MASP_TEST_SEED: &str = "NAMADA_MASP_TEST_SEED";
+
+/// The network to use for MASP
+#[cfg(feature = "mainnet")]
+const NETWORK: MainNetwork = MainNetwork;
+#[cfg(not(feature = "mainnet"))]
+const NETWORK: TestNetwork = TestNetwork;
+
+// TODO these could be exported from masp_proof crate
+/// Spend circuit name
+pub const SPEND_NAME: &str = "masp-spend.params";
+/// Output circuit name
+pub const OUTPUT_NAME: &str = "masp-output.params";
+/// Convert circuit name
+pub const CONVERT_NAME: &str = "masp-convert.params";
+
+lazy_static! {
+    /// MASP verifying keys load from parameters
+    static ref VERIFIYING_KEYS: PVKs =
+        {
+        let params_dir = get_params_dir();
+        let [spend_path, convert_path, output_path] =
+            [SPEND_NAME, CONVERT_NAME, OUTPUT_NAME].map(|p| params_dir.join(p));
+
+        #[cfg(feature = "download-params")]
+        if !spend_path.exists() || !convert_path.exists() || !output_path.exists() {
+            let paths = masp_proofs::download_masp_parameters(None).expect(
+                "MASP parameters were not present, expected the download to \
+                succeed",
+            );
+            if paths.spend != spend_path
+                || paths.convert != convert_path
+                || paths.output != output_path
+            {
+                panic!(
+                    "unrecoverable: downloaded missing masp params, but to an \
+                    unfamiliar path"
+                )
+            }
+        }
+        // size and blake2b checked here
+        let params = masp_proofs::load_parameters(
+            spend_path.as_path(),
+            output_path.as_path(),
+            convert_path.as_path(),
+        );
+        PVKs {
+            spend_vk: params.spend_vk,
+            convert_vk: params.convert_vk,
+            output_vk: params.output_vk
+        }
+    };
+}
+
+/// check_spend wrapper
+pub fn check_spend(
+    spend: &SpendDescription<<Authorized as Authorization>::SaplingAuth>,
+    sighash: &[u8; 32],
+    #[cfg(not(feature = "testing"))] ctx: &mut SaplingVerificationContext,
+    #[cfg(feature = "testing")]
+    ctx: &mut testing::MockSaplingVerificationContext,
+    parameters: &PreparedVerifyingKey<Bls12>,
+) -> bool {
+    let zkproof =
+        masp_proofs::bellman::groth16::Proof::read(spend.zkproof.as_slice());
+    let zkproof = match zkproof {
+        Ok(zkproof) => zkproof,
+        _ => return false,
+    };
+
+    ctx.check_spend(
+        spend.cv,
+        spend.anchor,
+        &spend.nullifier.0,
+        PublicKey(spend.rk.0),
+        sighash,
+        spend.spend_auth_sig,
+        zkproof,
+        parameters,
+    )
+}
+
+/// check_output wrapper
+pub fn check_output(
+    output: &OutputDescription<<<Authorized as Authorization>::SaplingAuth as masp_primitives::transaction::components::sapling::Authorization>::Proof>,
+    #[cfg(not(feature = "testing"))] ctx: &mut SaplingVerificationContext,
+    #[cfg(feature = "testing")]
+    ctx: &mut testing::MockSaplingVerificationContext,
+    parameters: &PreparedVerifyingKey<Bls12>,
+) -> bool {
+    let zkproof =
+        masp_proofs::bellman::groth16::Proof::read(output.zkproof.as_slice());
+    let zkproof = match zkproof {
+        Ok(zkproof) => zkproof,
+        _ => return false,
+    };
+    let epk =
+        masp_proofs::jubjub::ExtendedPoint::from_bytes(&output.ephemeral_key.0);
+    let epk = match epk.into() {
+        Some(p) => p,
+        None => return false,
+    };
+
+    ctx.check_output(output.cv, output.cmu, epk, zkproof, parameters)
+}
+
+/// check convert wrapper
+pub fn check_convert(
+    convert: &ConvertDescription<<<Authorized as Authorization>::SaplingAuth as masp_primitives::transaction::components::sapling::Authorization>::Proof>,
+    #[cfg(not(feature = "testing"))] ctx: &mut SaplingVerificationContext,
+    #[cfg(feature = "testing")]
+    ctx: &mut testing::MockSaplingVerificationContext,
+    parameters: &PreparedVerifyingKey<Bls12>,
+) -> bool {
+    let zkproof =
+        masp_proofs::bellman::groth16::Proof::read(convert.zkproof.as_slice());
+    let zkproof = match zkproof {
+        Ok(zkproof) => zkproof,
+        _ => return false,
+    };
+
+    ctx.check_convert(convert.cv, convert.anchor, zkproof, parameters)
+}
+
+/// Partially deauthorize the transparent bundle
+fn partial_deauthorize(
+    tx_data: &TransactionData<Authorized>,
+) -> Option<TransactionData<PartialAuthorized>> {
+    let transp = tx_data.transparent_bundle().and_then(|x| {
+        let mut tb = TransparentBuilder::empty();
+        for vin in &x.vin {
+            tb.add_input(TxOut {
+                asset_type: vin.asset_type,
+                value: vin.value,
+                address: vin.address,
+            })
+            .ok()?;
+        }
+        for vout in &x.vout {
+            tb.add_output(&vout.address, vout.asset_type, vout.value)
+                .ok()?;
+        }
+        tb.build()
+    });
+    if tx_data.transparent_bundle().is_some() != transp.is_some() {
+        return None;
+    }
+    Some(TransactionData::from_parts(
+        tx_data.version(),
+        tx_data.consensus_branch_id(),
+        tx_data.lock_time(),
+        tx_data.expiry_height(),
+        transp,
+        tx_data.sapling_bundle().cloned(),
+    ))
+}
+
+/// Verify a shielded transaction.
+pub fn verify_shielded_tx(transaction: &Transaction) -> bool {
+    tracing::info!("entered verify_shielded_tx()");
+
+    let sapling_bundle = if let Some(bundle) = transaction.sapling_bundle() {
+        bundle
+    } else {
+        return false;
+    };
+    let tx_data = transaction.deref();
+
+    // Partially deauthorize the transparent bundle
+    let unauth_tx_data = match partial_deauthorize(tx_data) {
+        Some(tx_data) => tx_data,
+        None => return false,
+    };
+
+    let txid_parts = unauth_tx_data.digest(TxIdDigester);
+    // the commitment being signed is shared across all Sapling inputs; once
+    // V4 transactions are deprecated this should just be the txid, but
+    // for now we need to continue to compute it here.
+    let sighash =
+        signature_hash(&unauth_tx_data, &SignableInput::Shielded, &txid_parts);
+
+    tracing::info!("sighash computed");
+
+    let PVKs {
+        spend_vk,
+        convert_vk,
+        output_vk,
+    } = load_pvks();
+
+    #[cfg(not(feature = "testing"))]
+    let mut ctx = SaplingVerificationContext::new(true);
+    #[cfg(feature = "testing")]
+    let mut ctx = testing::MockSaplingVerificationContext::new(true);
+    let spends_valid = sapling_bundle
+        .shielded_spends
+        .iter()
+        .all(|spend| check_spend(spend, sighash.as_ref(), &mut ctx, spend_vk));
+    let converts_valid = sapling_bundle
+        .shielded_converts
+        .iter()
+        .all(|convert| check_convert(convert, &mut ctx, convert_vk));
+    let outputs_valid = sapling_bundle
+        .shielded_outputs
+        .iter()
+        .all(|output| check_output(output, &mut ctx, output_vk));
+
+    if !(spends_valid && outputs_valid && converts_valid) {
+        return false;
+    }
+
+    tracing::info!("passed spend/output verification");
+
+    let assets_and_values: I128Sum = sapling_bundle.value_balance.clone();
+
+    tracing::info!(
+        "accumulated {} assets/values",
+        assets_and_values.components().len()
+    );
+
+    let result = ctx.final_check(
+        assets_and_values,
+        sighash.as_ref(),
+        sapling_bundle.authorization.binding_sig,
+    );
+    tracing::info!("final check result {result}");
+    result
+}
+
+mod tests {
+    /// quick and dirty test. will fail on size check
+    #[test]
+    #[should_panic(expected = "parameter file size is not correct")]
+    fn test_wrong_masp_params() {
+        use std::io::Write;
+
+        use super::{CONVERT_NAME, OUTPUT_NAME, SPEND_NAME};
+
+        let tempdir = tempfile::tempdir()
+            .expect("expected a temp dir")
+            .into_path();
+        let fake_params_paths =
+            [SPEND_NAME, OUTPUT_NAME, CONVERT_NAME].map(|p| tempdir.join(p));
+        for path in &fake_params_paths {
+            let mut f =
+                std::fs::File::create(path).expect("expected a temp file");
+            f.write_all(b"fake params")
+                .expect("expected a writable temp file");
+            f.sync_all()
+                .expect("expected a writable temp file (on sync)");
+        }
+
+        std::env::set_var(super::ENV_VAR_MASP_PARAMS_DIR, tempdir.as_os_str());
+        // should panic here
+        masp_proofs::load_parameters(
+            &fake_params_paths[0],
+            &fake_params_paths[1],
+            &fake_params_paths[2],
+        );
+    }
+
+    /// a more involved test, using dummy parameters with the right
+    /// size but the wrong hash.
+    #[test]
+    #[should_panic(expected = "parameter file is not correct")]
+    fn test_wrong_masp_params_hash() {
+        use masp_primitives::ff::PrimeField;
+        use masp_proofs::bellman::groth16::{
+            generate_random_parameters, Parameters,
+        };
+        use masp_proofs::bellman::{Circuit, ConstraintSystem, SynthesisError};
+        use masp_proofs::bls12_381::{Bls12, Scalar};
+
+        use super::{CONVERT_NAME, OUTPUT_NAME, SPEND_NAME};
+
+        struct FakeCircuit<E: PrimeField> {
+            x: E,
+        }
+
+        impl<E: PrimeField> Circuit<E> for FakeCircuit<E> {
+            fn synthesize<CS: ConstraintSystem<E>>(
+                self,
+                cs: &mut CS,
+            ) -> Result<(), SynthesisError> {
+                let x = cs.alloc(|| "x", || Ok(self.x)).unwrap();
+                cs.enforce(
+                    || {
+                        "this is an extra long constraint name so that rustfmt \
+                         is ok with wrapping the params of enforce()"
+                    },
+                    |lc| lc + x,
+                    |lc| lc + x,
+                    |lc| lc + x,
+                );
+                Ok(())
+            }
+        }
+
+        let dummy_circuit = FakeCircuit { x: Scalar::zero() };
+        let mut rng = rand::thread_rng();
+        let fake_params: Parameters<Bls12> =
+            generate_random_parameters(dummy_circuit, &mut rng)
+                .expect("expected to generate fake params");
+
+        let tempdir = tempfile::tempdir()
+            .expect("expected a temp dir")
+            .into_path();
+        // TODO: get masp to export these consts
+        let fake_params_paths = [
+            (SPEND_NAME, 49848572u64),
+            (OUTPUT_NAME, 16398620u64),
+            (CONVERT_NAME, 22570940u64),
+        ]
+        .map(|(p, s)| (tempdir.join(p), s));
+        for (path, size) in &fake_params_paths {
+            let mut f =
+                std::fs::File::create(path).expect("expected a temp file");
+            fake_params
+                .write(&mut f)
+                .expect("expected a writable temp file");
+            // the dummy circuit has one constraint, and therefore its
+            // params should always be smaller than the large masp
+            // circuit params. so this truncate extends the file, and
+            // extra bytes at the end do not make it invalid.
+            f.set_len(*size)
+                .expect("expected to truncate the temp file");
+            f.sync_all()
+                .expect("expected a writable temp file (on sync)");
+        }
+
+        std::env::set_var(super::ENV_VAR_MASP_PARAMS_DIR, tempdir.as_os_str());
+        // should panic here
+        masp_proofs::load_parameters(
+            &fake_params_paths[0].0,
+            &fake_params_paths[1].0,
+            &fake_params_paths[2].0,
+        );
+    }
+}
+
+#[cfg(any(test, feature = "testing"))]
+/// Tests and strategies for transactions
+pub mod testing {
+    use std::ops::AddAssign;
+    use std::sync::Mutex;
+
+    use bls12_381::{G1Affine, G2Affine};
+    use masp_primitives::consensus::testing::arb_height;
+    use masp_primitives::constants::SPENDING_KEY_GENERATOR;
+    use masp_primitives::ff::Field;
+    use masp_primitives::sapling::prover::TxProver;
+    use masp_primitives::sapling::redjubjub::Signature;
+    use masp_primitives::sapling::{ProofGenerationKey, Rseed};
+    use masp_primitives::transaction::builder::Builder;
+    use masp_primitives::transaction::components::GROTH_PROOF_SIZE;
+    use masp_proofs::bellman::groth16::Proof;
+    use proptest::prelude::*;
+    use proptest::sample::SizeRange;
+    use proptest::test_runner::TestRng;
+    use proptest::{collection, option, prop_compose};
+    use rand_core::CryptoRng;
+
+    use super::*;
+    use crate::address::testing::arb_address;
+    use crate::masp::types::{ShieldedTransfer, WalletMap};
+    use crate::masp::utils::find_valid_diversifier;
+    use crate::masp_primitives::consensus::BranchId;
+    use crate::masp_primitives::constants::VALUE_COMMITMENT_RANDOMNESS_GENERATOR;
+    use crate::masp_primitives::merkle_tree::FrozenCommitmentTree;
+    use crate::masp_primitives::sapling::keys::OutgoingViewingKey;
+    use crate::masp_primitives::sapling::redjubjub::PrivateKey;
+    use crate::masp_primitives::transaction::components::transparent::testing::arb_transparent_address;
+    use crate::masp_proofs::sapling::SaplingVerificationContextInner;
+    use crate::storage::testing::arb_epoch;
+    use crate::token::testing::arb_denomination;
+
+    /// A context object for verifying the Sapling components of a single Zcash
+    /// transaction. Same as SaplingVerificationContext, but always assumes the
+    /// proofs to be valid.
+    pub struct MockSaplingVerificationContext {
+        inner: SaplingVerificationContextInner,
+        zip216_enabled: bool,
+    }
+
+    impl MockSaplingVerificationContext {
+        /// Construct a new context to be used with a single transaction.
+        pub fn new(zip216_enabled: bool) -> Self {
+            MockSaplingVerificationContext {
+                inner: SaplingVerificationContextInner::new(),
+                zip216_enabled,
+            }
+        }
+
+        /// Perform consensus checks on a Sapling SpendDescription, while
+        /// accumulating its value commitment inside the context for later use.
+        #[allow(clippy::too_many_arguments)]
+        pub fn check_spend(
+            &mut self,
+            cv: jubjub::ExtendedPoint,
+            anchor: bls12_381::Scalar,
+            nullifier: &[u8; 32],
+            rk: PublicKey,
+            sighash_value: &[u8; 32],
+            spend_auth_sig: Signature,
+            zkproof: Proof<Bls12>,
+            _verifying_key: &PreparedVerifyingKey<Bls12>,
+        ) -> bool {
+            let zip216_enabled = true;
+            self.inner.check_spend(
+                cv,
+                anchor,
+                nullifier,
+                rk,
+                sighash_value,
+                spend_auth_sig,
+                zkproof,
+                &mut (),
+                |_, rk, msg, spend_auth_sig| {
+                    rk.verify_with_zip216(
+                        &msg,
+                        &spend_auth_sig,
+                        SPENDING_KEY_GENERATOR,
+                        zip216_enabled,
+                    )
+                },
+                |_, _proof, _public_inputs| true,
+            )
+        }
+
+        /// Perform consensus checks on a Sapling SpendDescription, while
+        /// accumulating its value commitment inside the context for later use.
+        #[allow(clippy::too_many_arguments)]
+        pub fn check_convert(
+            &mut self,
+            cv: jubjub::ExtendedPoint,
+            anchor: bls12_381::Scalar,
+            zkproof: Proof<Bls12>,
+            _verifying_key: &PreparedVerifyingKey<Bls12>,
+        ) -> bool {
+            self.inner.check_convert(
+                cv,
+                anchor,
+                zkproof,
+                &mut (),
+                |_, _proof, _public_inputs| true,
+            )
+        }
+
+        /// Perform consensus checks on a Sapling OutputDescription, while
+        /// accumulating its value commitment inside the context for later use.
+        pub fn check_output(
+            &mut self,
+            cv: jubjub::ExtendedPoint,
+            cmu: bls12_381::Scalar,
+            epk: jubjub::ExtendedPoint,
+            zkproof: Proof<Bls12>,
+            _verifying_key: &PreparedVerifyingKey<Bls12>,
+        ) -> bool {
+            self.inner.check_output(
+                cv,
+                cmu,
+                epk,
+                zkproof,
+                |_proof, _public_inputs| true,
+            )
+        }
+
+        /// Perform consensus checks on the valueBalance and bindingSig parts of
+        /// a Sapling transaction. All SpendDescriptions and
+        /// OutputDescriptions must have been checked before calling
+        /// this function.
+        pub fn final_check(
+            &self,
+            value_balance: I128Sum,
+            sighash_value: &[u8; 32],
+            binding_sig: Signature,
+        ) -> bool {
+            self.inner.final_check(
+                value_balance,
+                sighash_value,
+                binding_sig,
+                |bvk, msg, binding_sig| {
+                    bvk.verify_with_zip216(
+                        &msg,
+                        &binding_sig,
+                        VALUE_COMMITMENT_RANDOMNESS_GENERATOR,
+                        self.zip216_enabled,
+                    )
+                },
+            )
+        }
+    }
+
+    // This function computes `value` in the exponent of the value commitment
+    // base
+    fn masp_compute_value_balance(
+        asset_type: AssetType,
+        value: i128,
+    ) -> Option<jubjub::ExtendedPoint> {
+        // Compute the absolute value (failing if -i128::MAX is
+        // the value)
+        let abs = match value.checked_abs() {
+            Some(a) => a as u128,
+            None => return None,
+        };
+
+        // Is it negative? We'll have to negate later if so.
+        let is_negative = value.is_negative();
+
+        // Compute it in the exponent
+        let mut abs_bytes = [0u8; 32];
+        abs_bytes[0..16].copy_from_slice(&abs.to_le_bytes());
+        let mut value_balance = asset_type.value_commitment_generator()
+            * jubjub::Fr::from_bytes(&abs_bytes).unwrap();
+
+        // Negate if necessary
+        if is_negative {
+            value_balance = -value_balance;
+        }
+
+        // Convert to unknown order point
+        Some(value_balance.into())
+    }
+
+    // A context object for creating the Sapling components of a Zcash
+    // transaction.
+    pub struct SaplingProvingContext {
+        bsk: jubjub::Fr,
+        // (sum of the Spend value commitments) - (sum of the Output value
+        // commitments)
+        cv_sum: jubjub::ExtendedPoint,
+    }
+
+    // An implementation of TxProver that does everything except generating
+    // valid zero-knowledge proofs. Uses the supplied source of randomness to
+    // carry out its operations.
+    pub struct MockTxProver<R: RngCore>(pub Mutex<R>);
+
+    impl<R: RngCore> TxProver for MockTxProver<R> {
+        type SaplingProvingContext = SaplingProvingContext;
+
+        fn new_sapling_proving_context(&self) -> Self::SaplingProvingContext {
+            SaplingProvingContext {
+                bsk: jubjub::Fr::zero(),
+                cv_sum: jubjub::ExtendedPoint::identity(),
+            }
+        }
+
+        fn spend_proof(
+            &self,
+            ctx: &mut Self::SaplingProvingContext,
+            proof_generation_key: ProofGenerationKey,
+            _diversifier: Diversifier,
+            _rseed: Rseed,
+            ar: jubjub::Fr,
+            asset_type: AssetType,
+            value: u64,
+            _anchor: bls12_381::Scalar,
+            _merkle_path: MerklePath<Node>,
+        ) -> Result<
+            ([u8; GROTH_PROOF_SIZE], jubjub::ExtendedPoint, PublicKey),
+            (),
+        > {
+            // Initialize secure RNG
+            let mut rng = self.0.lock().unwrap();
+
+            // We create the randomness of the value commitment
+            let rcv = jubjub::Fr::random(&mut *rng);
+
+            // Accumulate the value commitment randomness in the context
+            {
+                let mut tmp = rcv;
+                tmp.add_assign(&ctx.bsk);
+
+                // Update the context
+                ctx.bsk = tmp;
+            }
+
+            // Construct the value commitment
+            let value_commitment = asset_type.value_commitment(value, rcv);
+
+            // This is the result of the re-randomization, we compute it for the
+            // caller
+            let rk = PublicKey(proof_generation_key.ak.into())
+                .randomize(ar, SPENDING_KEY_GENERATOR);
+
+            // Compute value commitment
+            let value_commitment: jubjub::ExtendedPoint =
+                value_commitment.commitment().into();
+
+            // Accumulate the value commitment in the context
+            ctx.cv_sum += value_commitment;
+
+            let mut zkproof = [0u8; GROTH_PROOF_SIZE];
+            let proof = Proof::<Bls12> {
+                a: G1Affine::generator(),
+                b: G2Affine::generator(),
+                c: G1Affine::generator(),
+            };
+            proof
+                .write(&mut zkproof[..])
+                .expect("should be able to serialize a proof");
+            Ok((zkproof, value_commitment, rk))
+        }
+
+        fn output_proof(
+            &self,
+            ctx: &mut Self::SaplingProvingContext,
+            _esk: jubjub::Fr,
+            _payment_address: masp_primitives::sapling::PaymentAddress,
+            _rcm: jubjub::Fr,
+            asset_type: AssetType,
+            value: u64,
+        ) -> ([u8; GROTH_PROOF_SIZE], jubjub::ExtendedPoint) {
+            // Initialize secure RNG
+            let mut rng = self.0.lock().unwrap();
+
+            // We construct ephemeral randomness for the value commitment. This
+            // randomness is not given back to the caller, but the synthetic
+            // blinding factor `bsk` is accumulated in the context.
+            let rcv = jubjub::Fr::random(&mut *rng);
+
+            // Accumulate the value commitment randomness in the context
+            {
+                let mut tmp = rcv.neg(); // Outputs subtract from the total.
+                tmp.add_assign(&ctx.bsk);
+
+                // Update the context
+                ctx.bsk = tmp;
+            }
+
+            // Construct the value commitment for the proof instance
+            let value_commitment = asset_type.value_commitment(value, rcv);
+
+            // Compute the actual value commitment
+            let value_commitment_point: jubjub::ExtendedPoint =
+                value_commitment.commitment().into();
+
+            // Accumulate the value commitment in the context. We do this to
+            // check internal consistency.
+            ctx.cv_sum -= value_commitment_point; // Outputs subtract from the total.
+
+            let mut zkproof = [0u8; GROTH_PROOF_SIZE];
+            let proof = Proof::<Bls12> {
+                a: G1Affine::generator(),
+                b: G2Affine::generator(),
+                c: G1Affine::generator(),
+            };
+            proof
+                .write(&mut zkproof[..])
+                .expect("should be able to serialize a proof");
+
+            (zkproof, value_commitment_point)
+        }
+
+        fn convert_proof(
+            &self,
+            ctx: &mut Self::SaplingProvingContext,
+            allowed_conversion: AllowedConversion,
+            value: u64,
+            _anchor: bls12_381::Scalar,
+            _merkle_path: MerklePath<Node>,
+        ) -> Result<([u8; GROTH_PROOF_SIZE], jubjub::ExtendedPoint), ()>
+        {
+            // Initialize secure RNG
+            let mut rng = self.0.lock().unwrap();
+
+            // We create the randomness of the value commitment
+            let rcv = jubjub::Fr::random(&mut *rng);
+
+            // Accumulate the value commitment randomness in the context
+            {
+                let mut tmp = rcv;
+                tmp.add_assign(&ctx.bsk);
+
+                // Update the context
+                ctx.bsk = tmp;
+            }
+
+            // Construct the value commitment
+            let value_commitment =
+                allowed_conversion.value_commitment(value, rcv);
+
+            // Compute value commitment
+            let value_commitment: jubjub::ExtendedPoint =
+                value_commitment.commitment().into();
+
+            // Accumulate the value commitment in the context
+            ctx.cv_sum += value_commitment;
+
+            let mut zkproof = [0u8; GROTH_PROOF_SIZE];
+            let proof = Proof::<Bls12> {
+                a: G1Affine::generator(),
+                b: G2Affine::generator(),
+                c: G1Affine::generator(),
+            };
+            proof
+                .write(&mut zkproof[..])
+                .expect("should be able to serialize a proof");
+
+            Ok((zkproof, value_commitment))
+        }
+
+        fn binding_sig(
+            &self,
+            ctx: &mut Self::SaplingProvingContext,
+            assets_and_values: &I128Sum,
+            sighash: &[u8; 32],
+        ) -> Result<Signature, ()> {
+            // Initialize secure RNG
+            let mut rng = self.0.lock().unwrap();
+
+            // Grab the current `bsk` from the context
+            let bsk = PrivateKey(ctx.bsk);
+
+            // Grab the `bvk` using DerivePublic.
+            let bvk = PublicKey::from_private(
+                &bsk,
+                VALUE_COMMITMENT_RANDOMNESS_GENERATOR,
+            );
+
+            // In order to check internal consistency, let's use the accumulated
+            // value commitments (as the verifier would) and apply
+            // value_balance to compare against our derived bvk.
+            {
+                let final_bvk = assets_and_values
+                    .components()
+                    .map(|(asset_type, value_balance)| {
+                        // Compute value balance for each asset
+                        // Error for bad value balances (-INT128_MAX value)
+                        masp_compute_value_balance(*asset_type, *value_balance)
+                    })
+                    .try_fold(ctx.cv_sum, |tmp, value_balance| {
+                        // Compute cv_sum minus sum of all value balances
+                        Result::<_, ()>::Ok(tmp - value_balance.ok_or(())?)
+                    })?;
+
+                // The result should be the same, unless the provided
+                // valueBalance is wrong.
+                if bvk.0 != final_bvk {
+                    return Err(());
+                }
+            }
+
+            // Construct signature message
+            let mut data_to_be_signed = [0u8; 64];
+            data_to_be_signed[0..32].copy_from_slice(&bvk.0.to_bytes());
+            data_to_be_signed[32..64].copy_from_slice(&sighash[..]);
+
+            // Sign
+            Ok(bsk.sign(
+                &data_to_be_signed,
+                &mut *rng,
+                VALUE_COMMITMENT_RANDOMNESS_GENERATOR,
+            ))
+        }
+    }
+
+    #[derive(Debug, Clone)]
+    // Adapts a CSPRNG from a PRNG for proptesting
+    pub struct TestCsprng<R: RngCore>(R);
+
+    impl<R: RngCore> CryptoRng for TestCsprng<R> {}
+
+    impl<R: RngCore> RngCore for TestCsprng<R> {
+        fn next_u32(&mut self) -> u32 {
+            self.0.next_u32()
+        }
+
+        fn next_u64(&mut self) -> u64 {
+            self.0.next_u64()
+        }
+
+        fn fill_bytes(&mut self, dest: &mut [u8]) {
+            self.0.fill_bytes(dest)
+        }
+
+        fn try_fill_bytes(
+            &mut self,
+            dest: &mut [u8],
+        ) -> Result<(), rand::Error> {
+            self.0.try_fill_bytes(dest)
+        }
+    }
+
+    prop_compose! {
+        // Expose a random number generator
+        pub fn arb_rng()(rng in Just(()).prop_perturb(|(), rng| rng)) -> TestRng {
+            rng
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary output description with the given value
+        pub fn arb_output_description(
+            asset_type: AssetType,
+            value: u64,
+        )(
+            mut rng in arb_rng().prop_map(TestCsprng),
+        ) -> (Option<OutgoingViewingKey>, masp_primitives::sapling::PaymentAddress, AssetType, u64, MemoBytes) {
+            let mut spending_key_seed = [0; 32];
+            rng.fill_bytes(&mut spending_key_seed);
+            let spending_key = masp_primitives::zip32::ExtendedSpendingKey::master(spending_key_seed.as_ref());
+
+            let viewing_key = ExtendedFullViewingKey::from(&spending_key).fvk.vk;
+            let (div, _g_d) = find_valid_diversifier(&mut rng);
+            let payment_addr = viewing_key
+                .to_payment_address(div)
+                .expect("a PaymentAddress");
+
+            (None, payment_addr, asset_type, value, MemoBytes::empty())
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary spend description with the given value
+        pub fn arb_spend_description(
+            asset_type: AssetType,
+            value: u64,
+        )(
+            address in arb_transparent_address(),
+            expiration_height in arb_height(BranchId::MASP, &TestNetwork),
+            mut rng in arb_rng().prop_map(TestCsprng),
+            prover_rng in arb_rng().prop_map(TestCsprng),
+        ) -> (ExtendedSpendingKey, Diversifier, Note, Node) {
+            let mut spending_key_seed = [0; 32];
+            rng.fill_bytes(&mut spending_key_seed);
+            let spending_key = masp_primitives::zip32::ExtendedSpendingKey::master(spending_key_seed.as_ref());
+
+            let viewing_key = ExtendedFullViewingKey::from(&spending_key).fvk.vk;
+            let (div, _g_d) = find_valid_diversifier(&mut rng);
+            let payment_addr = viewing_key
+                .to_payment_address(div)
+                .expect("a PaymentAddress");
+
+            let mut builder = Builder::<TestNetwork, _>::new_with_rng(
+                NETWORK,
+                // NOTE: this is going to add 20 more blocks to the actual
+                // expiration but there's no other exposed function that we could
+                // use from the masp crate to specify the expiration better
+                expiration_height.unwrap(),
+                rng,
+            );
+            // Add a transparent input to support our desired shielded output
+            builder.add_transparent_input(TxOut { asset_type, value, address }).unwrap();
+            // Finally add the shielded output that we need
+            builder.add_sapling_output(None, payment_addr, asset_type, value, MemoBytes::empty()).unwrap();
+            // Build a transaction in order to get its shielded outputs
+            let (transaction, metadata) = builder.build(
+                &MockTxProver(Mutex::new(prover_rng)),
+                &FeeRule::non_standard(U64Sum::zero()),
+            ).unwrap();
+            // Extract the shielded output from the transaction
+            let shielded_output = &transaction
+                .sapling_bundle()
+                .unwrap()
+                .shielded_outputs[metadata.output_index(0).unwrap()];
+
+            // Let's now decrypt the constructed notes
+            let (note, pa, _memo) = try_sapling_note_decryption::<_, OutputDescription<<<Authorized as Authorization>::SaplingAuth as masp_primitives::transaction::components::sapling::Authorization>::Proof>>(
+                &NETWORK,
+                1.into(),
+                &PreparedIncomingViewingKey::new(&viewing_key.ivk()),
+                shielded_output,
+            ).unwrap();
+            assert_eq!(payment_addr, pa);
+            // Make a path to out new note
+            let node = Node::new(shielded_output.cmu.to_repr());
+            (spending_key, div, note, node)
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary MASP denomination
+        pub fn arb_masp_digit_pos()(denom in 0..4u8) -> MaspDigitPos {
+            MaspDigitPos::from(denom)
+        }
+    }
+
+    // Maximum value for a note partition
+    const MAX_MONEY: u64 = 100;
+    // Maximum number of partitions for a note
+    const MAX_SPLITS: usize = 3;
+
+    prop_compose! {
+        // Arbitrarily partition the given vector of integers into sets and sum
+        // them
+        pub fn arb_partition(values: Vec<u64>)(buckets in ((!values.is_empty()) as usize)..=values.len())(
+            values in Just(values.clone()),
+            assigns in collection::vec(0..buckets, values.len()),
+            buckets in Just(buckets),
+        ) -> Vec<u64> {
+            let mut buckets = vec![0; buckets];
+            for (bucket, value) in assigns.iter().zip(values) {
+                buckets[*bucket] += value;
+            }
+            buckets
+        }
+    }
+
+    prop_compose! {
+        // Generate arbitrary spend descriptions with the given asset type
+        // partitioning the given values
+        pub fn arb_spend_descriptions(
+            asset: AssetData,
+            values: Vec<u64>,
+        )(partition in arb_partition(values))(
+            spend_description in partition
+                .iter()
+                .map(|value| arb_spend_description(
+                    encode_asset_type(
+                        asset.token.clone(),
+                        asset.denom,
+                        asset.position,
+                        asset.epoch,
+                    ).unwrap(),
+                    *value,
+                )).collect::<Vec<_>>()
+        ) -> Vec<(ExtendedSpendingKey, Diversifier, Note, Node)> {
+            spend_description
+        }
+    }
+
+    prop_compose! {
+        // Generate arbitrary output descriptions with the given asset type
+        // partitioning the given values
+        pub fn arb_output_descriptions(
+            asset: AssetData,
+            values: Vec<u64>,
+        )(partition in arb_partition(values))(
+            output_description in partition
+                .iter()
+                .map(|value| arb_output_description(
+                    encode_asset_type(
+                        asset.token.clone(),
+                        asset.denom,
+                        asset.position,
+                        asset.epoch,
+                    ).unwrap(),
+                    *value,
+                )).collect::<Vec<_>>()
+        ) -> Vec<(Option<OutgoingViewingKey>, masp_primitives::sapling::PaymentAddress, AssetType, u64, MemoBytes)> {
+            output_description
+        }
+    }
+
+    prop_compose! {
+        // Generate arbitrary spend descriptions with the given asset type
+        // partitioning the given values
+        pub fn arb_txouts(
+            asset: AssetData,
+            values: Vec<u64>,
+            address: TransparentAddress,
+        )(
+            partition in arb_partition(values),
+        ) -> Vec<TxOut> {
+            partition
+                .iter()
+                .map(|value| TxOut {
+                    asset_type: encode_asset_type(
+                        asset.token.clone(),
+                        asset.denom,
+                        asset.position,
+                        asset.epoch,
+                    ).unwrap(),
+                    value: *value,
+                    address,
+                }).collect::<Vec<_>>()
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary shielded MASP transaction builder
+        pub fn arb_shielded_builder(asset_range: impl Into<SizeRange>)(
+            assets in collection::hash_map(
+                arb_pre_asset_type(),
+                collection::vec(..MAX_MONEY, ..MAX_SPLITS),
+                asset_range,
+            ),
+        )(
+            expiration_height in arb_height(BranchId::MASP, &TestNetwork),
+            rng in arb_rng().prop_map(TestCsprng),
+            spend_descriptions in assets
+                .iter()
+                .map(|(asset, values)| arb_spend_descriptions(asset.clone(), values.clone()))
+                .collect::<Vec<_>>(),
+            output_descriptions in assets
+                .iter()
+                .map(|(asset, values)| arb_output_descriptions(asset.clone(), values.clone()))
+                .collect::<Vec<_>>(),
+            assets in Just(assets),
+        ) -> (
+            Builder::<TestNetwork, TestCsprng<TestRng>>,
+            HashMap<AssetData, u64>,
+        ) {
+            let mut builder = Builder::<TestNetwork, _>::new_with_rng(
+                NETWORK,
+                // NOTE: this is going to add 20 more blocks to the actual
+                // expiration but there's no other exposed function that we could
+                // use from the masp crate to specify the expiration better
+                expiration_height.unwrap(),
+                rng,
+            );
+            let mut leaves = Vec::new();
+            // First construct a Merkle tree containing all notes to be used
+            for (_esk, _div, _note, node) in spend_descriptions.iter().flatten() {
+                leaves.push(*node);
+            }
+            let tree = FrozenCommitmentTree::new(&leaves);
+            // Then use the notes knowing that they all have the same anchor
+            for (idx, (esk, div, note, _node)) in spend_descriptions.iter().flatten().enumerate() {
+                builder.add_sapling_spend(*esk, *div, *note, tree.path(idx)).unwrap();
+            }
+            for (ovk, payment_addr, asset_type, value, memo) in output_descriptions.into_iter().flatten() {
+                builder.add_sapling_output(ovk, payment_addr, asset_type, value, memo).unwrap();
+            }
+            (builder, assets.into_iter().map(|(k, v)| (k, v.iter().sum())).collect())
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary pre-asset type
+        pub fn arb_pre_asset_type()(
+            token in arb_address(),
+            denom in arb_denomination(),
+            position in arb_masp_digit_pos(),
+            epoch in option::of(arb_epoch()),
+        ) -> AssetData {
+            AssetData {
+                token,
+                denom,
+                position,
+                epoch,
+            }
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary shielding MASP transaction builder
+        pub fn arb_shielding_builder(
+            source: TransparentAddress,
+            asset_range: impl Into<SizeRange>,
+        )(
+            assets in collection::hash_map(
+                arb_pre_asset_type(),
+                collection::vec(..MAX_MONEY, ..MAX_SPLITS),
+                asset_range,
+            ),
+        )(
+            expiration_height in arb_height(BranchId::MASP, &TestNetwork),
+            rng in arb_rng().prop_map(TestCsprng),
+            txins in assets
+                .iter()
+                .map(|(asset, values)| arb_txouts(asset.clone(), values.clone(), source))
+                .collect::<Vec<_>>(),
+            output_descriptions in assets
+                .iter()
+                .map(|(asset, values)| arb_output_descriptions(asset.clone(), values.clone()))
+                .collect::<Vec<_>>(),
+            assets in Just(assets),
+        ) -> (
+            Builder::<TestNetwork, TestCsprng<TestRng>>,
+            HashMap<AssetData, u64>,
+        ) {
+            let mut builder = Builder::<TestNetwork, _>::new_with_rng(
+                NETWORK,
+                // NOTE: this is going to add 20 more blocks to the actual
+                // expiration but there's no other exposed function that we could
+                // use from the masp crate to specify the expiration better
+                expiration_height.unwrap(),
+                rng,
+            );
+            for txin in txins.into_iter().flatten() {
+                builder.add_transparent_input(txin).unwrap();
+            }
+            for (ovk, payment_addr, asset_type, value, memo) in output_descriptions.into_iter().flatten() {
+                builder.add_sapling_output(ovk, payment_addr, asset_type, value, memo).unwrap();
+            }
+            (builder, assets.into_iter().map(|(k, v)| (k, v.iter().sum())).collect())
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary deshielding MASP transaction builder
+        pub fn arb_deshielding_builder(
+            target: TransparentAddress,
+            asset_range: impl Into<SizeRange>,
+        )(
+            assets in collection::hash_map(
+                arb_pre_asset_type(),
+                collection::vec(..MAX_MONEY, ..MAX_SPLITS),
+                asset_range,
+            ),
+        )(
+            expiration_height in arb_height(BranchId::MASP, &TestNetwork),
+            rng in arb_rng().prop_map(TestCsprng),
+            spend_descriptions in assets
+                .iter()
+                .map(|(asset, values)| arb_spend_descriptions(asset.clone(), values.clone()))
+                .collect::<Vec<_>>(),
+            txouts in assets
+                .iter()
+                .map(|(asset, values)| arb_txouts(asset.clone(), values.clone(), target))
+                .collect::<Vec<_>>(),
+            assets in Just(assets),
+        ) -> (
+            Builder::<TestNetwork, TestCsprng<TestRng>>,
+            HashMap<AssetData, u64>,
+        ) {
+            let mut builder = Builder::<TestNetwork, _>::new_with_rng(
+                NETWORK,
+                // NOTE: this is going to add 20 more blocks to the actual
+                // expiration but there's no other exposed function that we could
+                // use from the masp crate to specify the expiration better
+                expiration_height.unwrap(),
+                rng,
+            );
+            let mut leaves = Vec::new();
+            // First construct a Merkle tree containing all notes to be used
+            for (_esk, _div, _note, node) in spend_descriptions.iter().flatten() {
+                leaves.push(*node);
+            }
+            let tree = FrozenCommitmentTree::new(&leaves);
+            // Then use the notes knowing that they all have the same anchor
+            for (idx, (esk, div, note, _node)) in spend_descriptions.into_iter().flatten().enumerate() {
+                builder.add_sapling_spend(esk, div, note, tree.path(idx)).unwrap();
+            }
+            for txout in txouts.into_iter().flatten() {
+                builder.add_transparent_output(&txout.address, txout.asset_type, txout.value).unwrap();
+            }
+            (builder, assets.into_iter().map(|(k, v)| (k, v.iter().sum())).collect())
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary MASP shielded transfer
+        pub fn arb_shielded_transfer(
+            asset_range: impl Into<SizeRange>,
+        )(asset_range in Just(asset_range.into()))(
+            (builder, asset_types) in arb_shielded_builder(asset_range),
+            epoch in arb_epoch(),
+            rng in arb_rng().prop_map(TestCsprng),
+        ) -> (ShieldedTransfer, HashMap<AssetData, u64>) {
+            let (masp_tx, metadata) = builder.clone().build(
+                &MockTxProver(Mutex::new(rng)),
+                &FeeRule::non_standard(U64Sum::zero()),
+            ).unwrap();
+            (ShieldedTransfer {
+                builder: builder.map_builder(WalletMap),
+                metadata,
+                masp_tx,
+                epoch,
+            }, asset_types)
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary MASP shielded transfer
+        pub fn arb_shielding_transfer(
+            source: TransparentAddress,
+            asset_range: impl Into<SizeRange>,
+        )(asset_range in Just(asset_range.into()))(
+            (builder, asset_types) in arb_shielding_builder(
+                source,
+                asset_range,
+            ),
+            epoch in arb_epoch(),
+            rng in arb_rng().prop_map(TestCsprng),
+        ) -> (ShieldedTransfer, HashMap<AssetData, u64>) {
+            let (masp_tx, metadata) = builder.clone().build(
+                &MockTxProver(Mutex::new(rng)),
+                &FeeRule::non_standard(U64Sum::zero()),
+            ).unwrap();
+            (ShieldedTransfer {
+                builder: builder.map_builder(WalletMap),
+                metadata,
+                masp_tx,
+                epoch,
+            }, asset_types)
+        }
+    }
+
+    prop_compose! {
+        // Generate an arbitrary MASP shielded transfer
+        pub fn arb_deshielding_transfer(
+            target: TransparentAddress,
+            asset_range: impl Into<SizeRange>,
+        )(asset_range in Just(asset_range.into()))(
+            (builder, asset_types) in arb_deshielding_builder(
+                target,
+                asset_range,
+            ),
+            epoch in arb_epoch(),
+            rng in arb_rng().prop_map(TestCsprng),
+        ) -> (ShieldedTransfer, HashMap<AssetData, u64>) {
+            let (masp_tx, metadata) = builder.clone().build(
+                &MockTxProver(Mutex::new(rng)),
+                &FeeRule::non_standard(U64Sum::zero()),
+            ).unwrap();
+            (ShieldedTransfer {
+                builder: builder.map_builder(WalletMap),
+                metadata,
+                masp_tx,
+                epoch,
+            }, asset_types)
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+/// Implementation of MASP functionality depending on a standard filesystem
+pub mod fs {
+    use std::fs::{File, OpenOptions};
+    use std::io::{Read, Write};
+
+    use super::*;
+    use crate::masp::shielded_ctx::ShieldedContext;
+    use crate::masp::types::ContextSyncStatus;
+    use crate::masp::utils::ShieldedUtils;
+
+    /// Shielded context file name
+    const FILE_NAME: &str = "shielded.dat";
+    const TMP_FILE_NAME: &str = "shielded.tmp";
+    const SPECULATIVE_FILE_NAME: &str = "speculative_shielded.dat";
+    const SPECULATIVE_TMP_FILE_NAME: &str = "speculative_shielded.tmp";
+
+    #[derive(Debug, BorshSerialize, BorshDeserialize, Clone)]
+    /// An implementation of ShieldedUtils for standard filesystems
+    pub struct FsShieldedUtils {
+        #[borsh(skip)]
+        context_dir: PathBuf,
+    }
+
+    impl FsShieldedUtils {
+        /// Initialize a shielded transaction context that identifies notes
+        /// decryptable by any viewing key in the given set
+        pub fn new(context_dir: PathBuf) -> ShieldedContext<Self> {
+            // Make sure that MASP parameters are downloaded to enable MASP
+            // transaction building and verification later on
+            let params_dir = get_params_dir();
+            let spend_path = params_dir.join(SPEND_NAME);
+            let convert_path = params_dir.join(CONVERT_NAME);
+            let output_path = params_dir.join(OUTPUT_NAME);
+            if !(spend_path.exists()
+                && convert_path.exists()
+                && output_path.exists())
+            {
+                println!("MASP parameters not present, downloading...");
+                masp_proofs::download_masp_parameters(None)
+                    .expect("MASP parameters not present or downloadable");
+                println!(
+                    "MASP parameter download complete, resuming execution..."
+                );
+            }
+            // Finally initialize a shielded context with the supplied directory
+
+            let sync_status =
+                if std::fs::read(context_dir.join(SPECULATIVE_FILE_NAME))
+                    .is_ok()
+                {
+                    // Load speculative state
+                    ContextSyncStatus::Speculative
+                } else {
+                    ContextSyncStatus::Confirmed
+                };
+
+            let utils = Self { context_dir };
+            ShieldedContext {
+                utils,
+                sync_status,
+                ..Default::default()
+            }
+        }
+    }
+
+    impl Default for FsShieldedUtils {
+        fn default() -> Self {
+            Self {
+                context_dir: PathBuf::from(FILE_NAME),
+            }
+        }
+    }
+
+    #[cfg_attr(feature = "async-send", async_trait::async_trait)]
+    #[cfg_attr(not(feature = "async-send"), async_trait::async_trait(?Send))]
+    impl ShieldedUtils for FsShieldedUtils {
+        fn local_tx_prover(&self) -> LocalTxProver {
+            if let Ok(params_dir) = env::var(ENV_VAR_MASP_PARAMS_DIR) {
+                let params_dir = PathBuf::from(params_dir);
+                let spend_path = params_dir.join(SPEND_NAME);
+                let convert_path = params_dir.join(CONVERT_NAME);
+                let output_path = params_dir.join(OUTPUT_NAME);
+                LocalTxProver::new(&spend_path, &output_path, &convert_path)
+            } else {
+                LocalTxProver::with_default_location()
+                    .expect("unable to load MASP Parameters")
+            }
+        }
+
+        /// Try to load the last saved shielded context from the given context
+        /// directory. If this fails, then leave the current context unchanged.
+        async fn load<U: ShieldedUtils + MaybeSend>(
+            &self,
+            ctx: &mut ShieldedContext<U>,
+            force_confirmed: bool,
+        ) -> std::io::Result<()> {
+            // Try to load shielded context from file
+            let file_name = if force_confirmed {
+                FILE_NAME
+            } else {
+                match ctx.sync_status {
+                    ContextSyncStatus::Confirmed => FILE_NAME,
+                    ContextSyncStatus::Speculative => SPECULATIVE_FILE_NAME,
+                }
+            };
+            let mut ctx_file = File::open(self.context_dir.join(file_name))?;
+            let mut bytes = Vec::new();
+            ctx_file.read_to_end(&mut bytes)?;
+            // Fill the supplied context with the deserialized object
+            *ctx = ShieldedContext {
+                utils: ctx.utils.clone(),
+                ..ShieldedContext::<U>::deserialize(&mut &bytes[..])?
+            };
+            Ok(())
+        }
+
+        /// Save this confirmed shielded context into its associated context
+        /// directory. At the same time, delete the speculative file if present
+        async fn save<U: ShieldedUtils + MaybeSync>(
+            &self,
+            ctx: &ShieldedContext<U>,
+        ) -> std::io::Result<()> {
+            // TODO: use mktemp crate?
+            let (tmp_file_name, file_name) = match ctx.sync_status {
+                ContextSyncStatus::Confirmed => (TMP_FILE_NAME, FILE_NAME),
+                ContextSyncStatus::Speculative => {
+                    (SPECULATIVE_TMP_FILE_NAME, SPECULATIVE_FILE_NAME)
+                }
+            };
+            let tmp_path = self.context_dir.join(tmp_file_name);
+            {
+                // First serialize the shielded context into a temporary file.
+                // Inability to create this file implies a simultaneuous write
+                // is in progress. In this case, immediately
+                // fail. This is unproblematic because the data
+                // intended to be stored can always be re-fetched
+                // from the blockchain.
+                let mut ctx_file = OpenOptions::new()
+                    .write(true)
+                    .create_new(true)
+                    .open(tmp_path.clone())?;
+                let mut bytes = Vec::new();
+                ctx.serialize(&mut bytes)
+                    .expect("cannot serialize shielded context");
+                ctx_file.write_all(&bytes[..])?;
+            }
+            // Atomically update the old shielded context file with new data.
+            // Atomicity is required to prevent other client instances from
+            // reading corrupt data.
+            std::fs::rename(tmp_path, self.context_dir.join(file_name))?;
+
+            // Remove the speculative file if present since it's state is
+            // overruled by the confirmed one we just saved
+            if let ContextSyncStatus::Confirmed = ctx.sync_status {
+                let _ = std::fs::remove_file(
+                    self.context_dir.join(SPECULATIVE_FILE_NAME),
+                );
+            }
+
+            Ok(())
+        }
+    }
+}
