+++ crates/sdk/src/queries/mod.rs
@@ -427,7 +427,9 @@ pub trait Client {
 
 #[cfg_attr(feature = "async-send", async_trait::async_trait)]
 #[cfg_attr(not(feature = "async-send"), async_trait::async_trait(?Send))]
-impl<C: tendermint_rpc::client::Client + std::marker::Sync> Client for C {
+impl<C: tendermint_rpc::client::Client + std::marker::Sync + Clone> Client
+    for C
+{
     type Error = Error;
 
     async fn request(
