+++ crates/sdk/src/queries/shell.rs
@@ -30,7 +30,7 @@ use crate::events::{Event, EventType};
 use crate::ibc::core::host::types::identifiers::{
     ChannelId, ClientId, PortId, Sequence,
 };
-use crate::masp::MaspTokenRewardData;
+use crate::masp::types::MaspTokenRewardData;
 use crate::queries::types::{RequestCtx, RequestQuery};
 use crate::queries::{require_latest_height, EncodedResponseQuery};
 use crate::tendermint::merkle::proof::ProofOps;
