+++ crates/sdk/src/rpc.rs
@@ -45,7 +45,7 @@ use crate::error::{EncodingError, Error, QueryError, TxSubmitError};
 use crate::events::Event;
 use crate::internal_macros::echo_error;
 use crate::io::Io;
-use crate::masp::MaspTokenRewardData;
+use crate::masp::types::MaspTokenRewardData;
 use crate::queries::vp::pos::EnrichedBondsAndUnbondsDetails;
 use crate::queries::{Client, RPC};
 use crate::tendermint::block::Height;
