+++ crates/sdk/src/tx.rs
@@ -56,8 +56,9 @@ use crate::args::{self, InputAmount};
 use crate::control_flow::time;
 use crate::error::{EncodingError, Error, QueryError, Result, TxSubmitError};
 use crate::io::Io;
-use crate::masp::TransferErr::Build;
-use crate::masp::{ShieldedContext, ShieldedTransfer};
+use crate::masp::types::ShieldedTransfer;
+use crate::masp::types::TransferErr::Build;
+use crate::masp::ShieldedContext;
 use crate::queries::Client;
 use crate::rpc::{
     self, query_wasm_code_hash, validate_amount, InnerTxResult,
